Introduce config prop overrideDefaultSearchParams, defaulting to no overrides
