+++ src/components/Search/createSearchQuery.tsx
@@ -1,5 +1,5 @@
 import { ProjectConfig } from "../../model/ProjectConfig.ts";
-import { FacetEntry, NamedFacetAgg } from "../../model/Search.ts";
+import { FacetEntry, NamedFacetAgg, SearchQuery } from "../../model/Search.ts";
 import _ from "lodash";
 import { blankSearchQuery } from "../../stores/search/default-query-slice.ts";
 
@@ -7,7 +7,7 @@ export function createSearchQuery(props: {
   projectConfig: ProjectConfig;
   aggs?: NamedFacetAgg[];
   dateFacets?: FacetEntry[];
-}) {
+}): SearchQuery {
   const { projectConfig, aggs, dateFacets } = props;
 
   const configuredSearchQuery = {
